Added vision to the agents
I added 2 new arrays (spotsSeeing for the spots the agent can currently see, and spotsSeen for all spots seen from the agent) to be able to identify another agent with and pursue accordingly. I've also added a time sleep function to draw out the simulation and have a more friendly viewing experience of what's going on. There is keyboard functionality for Windows systems from the commented msvcrt code, but I plan to adapt a different system soon.