Improved agent vision
The agents can now see fully down the vertical halls (2 spaces) as well as see around a vertical hall if they're next to a corner. I've also improved the way the seeker goes to the hider if the hider is in sight. I haven't implemented the horizontal corner look around yet but will soon.