Sight improved
The vertical sight was fixed along with implementation of horizontal corner viewing. Now if an agent is around the corner then they can view what's down the hallways of said corner without having to stand in the middle of it. The horizontal viewing could use two extra blocks, but that would allow them to see pathways while behind a corner which isn't yet put up in vertical viewing (we could do this by just adding one extra viewing block to the updateTop/updateBottom functions and uncommenting lines 386,387,388, 397,398,399, 428,429,430 and 439,440,441[so the horizontal has one extra viewing sight])