Adding more sensory data
The agent class now has it's own maze string that it keeps track of and updates which gives us more sensory data to work with. Given the memory of the layout of the maze for each agent, they can determine the best paths to take when pursuing/escaping each other. I've added the seekers maze into the print out to more give us more visual aid when working with the data.