Updated sights for debugging
I've updated the sight after the agent moves so when we print it out it matches where the agent is as opposed to where he was before. I've also made it print out the agent maze first now because it's easier to track the real maze in a smaller window while it's running.