Uniform cost path
Uniform cost path is now implemented, right now I just have it being fed the real maze string as opposed to the agent just to demonstrate it works. I'm going to implement a traversal function next that have it behave the way we want where the agent doesn't know where the hider is immediately and will use some more visual clues in it to help the agent.