Uniform cost seeking with traversal function
Added the uniform cost path to the seeking agent once the hider is found, and the seeker performs the random traversal that we've implemented before until then. Going to improve the agents visual functionality next by say having it look down hallways and decide to traverse them if there's openings to other paths.