Machine Learning Implementation
Machine learning has been implemented by applying the features of 'Walls Around Agent', 'Steps From Entrance', and 'Path Decision Choices From Entrance' to predict the labeling 'Steps Taken'. If the hider knows of a certain number of available spaces depending on the round or knows 90% of available spaces, then the hider will begin to decide which space is the best from the spaces that they know about and begin traversing their way to that destination. The game goes through 5 rounds and prints the steps taken according to each round which tends to go up as the rounds progress as the hider makes better hiding choices. The hider uses data from the program as well as data from the csv file 'data-example.csv' to build, fit, and predict its model from.